Remove Clojure's special completion
It wasn't working properly, because currently the framework files
(autoload/test/**) didn't get loaded, so the completion was always
"file". Even if we did manage to load them, that means that we would
effectively load the whole plugin at startup, which could hurt the
startup time.

And it's not necessary, we can just translate filenames into namespaces
(since they have a 1-1 corresponding anyway, as I understood).