Add neoterm strategy
Neoterm provides better customization options for how terminals are spawned in
neovim. It behaves more like dispatch in that it can send the command to a new
split. Neoterm has some built in test commands but they aren't as good as the
ones provided by vim-test. This strategy allows you to get the best of both
worlds.