Don't try to predict needed compilers for JS
Detecting CoffeeScript is error-prone, since there may be a CoffeeScript
file somewhere in node_modules/, but that module can still ship with the
compiled JavaScript files, meaning that the CoffeeScript compiler isn't
necessary and the user doesn't have to have it installed. Mocha
autoloads a test/mocha.opts file, where users can specify their compiler
options (they should do this anyway, so they don't have to specify it on
the command line).

The support for JSX files probably isn't that error-prone, since the
user who has this files most likely is working with them directly. But
we remove the compiler options for these as well for consistency and to
avoid any potential problems in the future (maybe the JSX compiler will
be extracted out of Babel).

Fixes #52