Remove the Vagrant strategy
This strategy was supposed to be a transformation in the first place, so
it shouldn't be a strategy anyway. The reason why we're removing it
completely is because we want to keep writing transformations out of
scope of vim-test, since implementations will likely be tied to tools
which interface changes often, so maintaining them would most likely be
a pain.  It's much better to leave writing transformations to the users.