Allow :TestLast to accept arguments
It can often happen that, when we're rerunning the last test, we want to
also add some extra options. For example, if the test is failing, we may
want to rerun it, but this time tell the test runner to add some extra
verbosity. So we allow passing options to the :TestLast command:

:TestLast --backtrace

The extra arguments added to :TestLast are saved with the last command,
so the next time we run :TestLast, the "--backtrace" option will already
be included.