Require the argument in test#run_last()
We made the argument optional to preserve backwards compatibility in
case anyone was calling test#run_last() directly. However, we already
broke backwards compatibility in test#run() by switching from a string
"arguments" to an array, so we might as well sacrifice the backwards
compatibility in test#run_last() for better readability.