Improve the Neovim strategy
Previously the Neovim strategy would open a Terminal window, and after
exiting it would switch to whatever buffer was the closest to the
Terminal window, instead of switching to back to the original window.

So we rewrite the strategy to act more like Dispatch.vim. Running tests
now doesn't lose focus from the current buffer. If tests pass, the
terminal window is automatically closed, otherwise it's left open so
that failures can be inspected.

Fixes #98