Restore neovim strategy's previous behaviour
The neovim strategy now focuses on the terminal window in insert mode.
If the user wants to move inside the Terminal buffer, they can use
<C-\><C-n> to get to normal mode (they can map it to <Esc> via
:tnoremap).

Previously when the terminal window closed, Vim would focus on the
nearest buffer instead of the buffer that was focused when the terminal
was opened. We fix that by adding a buffer-local autocommand which
returns to the previous window.

Fixes #118