FIX better support for options in maventest
maventest.vim could not support the vim-test way of adding options for a test
runner. For example if a user tried to pass `-f /path/to/pom.xml` as an option
parameter as follows:

`let test#java#maventest#options = '-f /path/to/pom.xml'`

the script would end up calling

`mvn test -Dtest=-f /path/to/pom.xml TestFileName`

which cannot work. Corrected the argument handling to support any number of
option parameters. The assumption is that 0 to 1 test names are passed to the
plugin. The example above now results to:

`mvn test -f /path/to/pom.xml -Dtest=TestFileName`