On Neovim use the "neovim" strategy by default
The "basic" strategy sucks on Neovim, because while on Vim you get
switched back to the Terminal (and can scroll in the output), in Neovim
shell commands are executed inside of Neovim, with colors showing up as
terminal escape characters.

Since I don't see any advantages of the "basic" strategy compared to the
"neovim" strategy, we make the "neovim" strategy the default on Neovim.