Open full-screen terminal on Neovim by default
On Vim the basic strategy switches to the terminal, runs the command,
allowing you to scroll through the colored output. When you want to come
back to your Vim session you can just press <Enter>.

On Neovim the ":!" command looks quite different. It runs the command
from within the Vim session on the bottom, without any colors in the
output (if there are colors, they will be shown as raw ANSI escape
sequences). You cannot scroll through the output, you can just press
<Enter> to go back to editing.

I think we can safely assume that nobody wants ":!" to be used in Neovim
for running tests, so we modify the basic strategy to open a terminal
session in a new tab, imitating the ":!" semantics.

I also found out that if you run the test command by editing a
"term://..." file, the terminal buffer will be automatically deleted
after the user preses <Enter> from the insert mode, so we were able to
remove the additional logic for deleting the terminal buffer.

Closes #203