Remove code for manually deleting the terminal buffer
It's not necessary to manually delete the terminal buffer, because it
will be automatically deleted once the user presses "Enter" from the
insert mode. This was added before when terminal session was left in
normal mode, which would compel the user to close the session with :q,
in which case the buffer would still remain open in the background.