Update Change Directory Logic When Running Tests
I'm having an issue with my setup where I'm using `vtr` dispatch strategy and I'm targeting
a tmux pane that runs a docker container. My project path in the docker container is different from the one in my host machine. So whenever I'm executing a test, there is `cd getcwd()` that will
be called and it resulted with `no such file or directory`.

Here I'm moving the code related to changing directory based on defined project root to the strategy file.

I've added a function that determine which project directory to change to with following priority

1. project_root
2. cwd (current working directory)