Option to Set Secondary Project Root for Docker Container
I'm having an issue with my setup where I'm using `vtr` dispatch strategy and I'm targeting
a tmux pane that runs a docker container. My project path in the docker container is different from the one in my host machine. So whenever I'm executing a test, there is `cd getcwd()` that will
be called and it resulted with `no such file or directory`.

Here i'm introducing `test#secondary_project_root` setting which will be useful for a setup
that runs in docker environment which different project path.

I've added a function that determine which project directory to change to with following priority

1. secondary_project_root
2. project_root
3. cwd (current working directory)