[Option 3] Remove Change Directory in Strategy
**Why we need this change?**

I'm having an issue with my setup where I'm using vtr dispatch strategy and I'm targeting
a tmux pane that runs a docker container. My project path in the docker container is different from the one in my host machine. So whenever I'm executing a test, there is cd getcwd() that will
be called and it resulted with no such file or directory.

**What has changed?**

I've removed the related code for changing to test directory in strategy file.