Fix reloading .vimrc by introducing test#custom_runners
It's not right that "test#runners" is used both for setting custom
runners and later vim-test extends it with default runners. This causes
issues with reloading .vimrc when custom runners are used.

So instead users should use the new "test#custom_runners" variable, and
this time the list of runners will be dynamically generated, which means
it will work better with .vimrc reloading.

We still keep backwards compatibility with setting "test#runners".

Fixes #306