Get the current file path before `before_run`.
When `test#project_root` is used `before_run` has a side effect of
changing `expand('%')` from a relative path to an absolute path. I
suspect this is because of the `execute 'cd'` call.

This side effect can prevent tests from running if the test command is
dependent on the relative path.

For instance, I used the Vagrant transformation listed in the README.
Because my Vagrant VM is in a different directory, I had to add
`test#project_root` to make the `vagrant ssh` call work. With the side
effect occurring, the test in the vagrant ssh connection would fail
because the position['file'] path was set to the absolute path to the
test file *on my host machine*.

I also had to move the first `get_position` call because `get_position`
tries to compare `a:path` to `expand('%')` to pick the line and column
numbers. If `get_position` is called after `before_run`, then `a:path`
does not match as it should.