Add vertical neovim strategy
Add Vertical Neovim strategy to run tests in a terminal in a vertical, isntead of horiontal.
 
Sorry for this lousy PR, but before I improve it I'd like to know if you are accepting this kind of strategies at the repository. Other options would be to just document this kind of strategy for the user to add it to their init.vim or to add options to the basic `neovim` strategy. Thad to you think?