vimterminal: ensure only one window is shown at a time
This change stores the number of the buffer containing the test output.
On subsequent runs, we then close the existing buffer/window (if any)
before creating a new buffer/window for the new run. Visually, this
creates the effect that the existing window is 'replaced' for the new
run.

I've also added a TestClose command which closes the buffer, which users
could create their own mapping for.