vimterminal: use Ctrl-C to close the terminal buffer
I set about using tnoremap to make it possible to close the terminal
while the command is still running. However, I don't think <Enter> is a
good mapping for this, since the terminal might actually need to receive
an <Enter> key. I think <Ctrl-C> is a natural mapping, and using it
allows us to have the same mapping both when the command is running and
when it has finished. I've also changed the mapping to use q! which will
kill the job if it's still running (which I think is the desirable
behaviour in the use case of stopping a running test).