vimterminal: switch back to previous window
Since I've changed the vimterminal strategy to reuse existing windows,
and added a :TestClose command to enable closing the window from outside
of it, I think it is desirable to not stay within the test runner
window while the test is running.

If we stay within it, most key combinations we might press will be sent
to the terminal while the command is still running. You have to use an
obscure key combination (Ctrl-w Shift-N) to prevent this.

If we move outside the window, we can continue to edit/read our code
while the test is running asynchronously.