PHPUnit - Handles @test annotations
The @test annotation was not handle.
Event if there were tests pretending it was actually only looking for
the first public method it could find above the cursor.

It was not a problem until I had a test with an anonymous class inside
it, see the fixture for PHPUnit for an example.
In this case, the test#base#nearest_test will stop at the first public
method inside the anonymous class and break the functionality.

What I did for PHPUnit is to use the standard test#base#nearest_test to
look for the first public method starting with 'test' or the first @test
docblock.
If it's the docblock which is found then we look forward from the
docblock position for the first public method we can find.
It should be the method to which the docblock refers to.