Fix options being appended in delve runner
# What
Insert a `--` in the dlv test command between the package name and where
test options are supplied.

# Why
The `--` is necessary to tell the dlv test command where options for go
test can be supplied vs where the package list start and ends.

When I originally developed the delve runner I had included a `--` and
then removed it thinking it was unnecessary just before I opened the PR
(#385). The test scenarios I'd coded didn't pickup on the options now
being ignored.