Replace rails check with railties check
Some applications don't use the `rails` gem explicitly and instead
depend on its components separately, such as `activerecord`,
`activesupport`, etc. When this is how an application requires rails, it
causes the test command to use `bin/rake` instead of `bin/rails` causing
tests not to run.

This resolves the issue by checking the `railties` version which should
always be present, even when `rails` isn't.