Return to the same directory on failed run
With this change, if the failure "Not a test file" is encountered, `after_run` is ran before returning. `after_run` will run `cd -` which will resolve to the directory before the last `cd`. This achieves some symmetry if `g:test#project_root` is set(cd to the directory before tests are run => cd back).