Support Go +build tags (#641)
When current file's got build tag constraints, by default it may not be
compiled by `go test`. This commit tries to parse the +build tags [1]
and add `-tags=foo,bar` to the command line.

The new go:build [2] is not supported since it's quite troublesome to
process. Users of Go 1.17+ can have both go:build and +build in a file
as long as they are consistent, or manually `:TestFile -tags=foo,bar`.

[1] https://pkg.go.dev/cmd/go@go1.16.15#hdr-Build_constraints
[2] https://pkg.go.dev/cmd/go@go1.18#hdr-Build_constraints