Simplify test pattern customization for CommonTest runner
The g:test#erlang#patterns used only by CommonTest is replaced with
g:test#erlang#commontest#test_pattern.  This is BC break, although the
variables are not documented.

In the Common Test framework, a test case function can have any name.
Some naming conventions exist though.  For simple test suites with no
groups these conventions can be used to run :TestNearest.  So it may be
necessary to customize the test pattern.

Neither the Common Test framework nor Erlang have a concept equivalent
to a namespace.  So the pattern for :TestNearest can be just a string.
This way is also used for EUnit and PropEr.