Fix triggering of autocmds when exiting terminal
When using vimterminal or neovim strategy a BufDelete autocmd is used to
jump back to the previous window, but this is not compatible with other
plugins or configuration that relies on BufEnter or WinEnter autocmds as
they do not get triggered. Fix by triggering nested autocmds.

For example, this broke lightline such that it would appear that the
current window was actually the previous window as the lightline update
autocmds would not be triggered. There are also examples in the vim docs
that would be broken without using ++nested, e.g. :help 'cursorcolumn'